--- src/components/Modal/styles.scss
@@ -1,1 +1,0 @@
-@import "../../styles/components/modal";
