+++ src/components/ModalMyList/index.js
@@ -1,19 +1,15 @@
-import React, { useEffect, useState } from "react";
+import React from "react";
 import { Formik, Form, Field, ErrorMessage } from "formik";
 import PropTypes from "prop-types";
 import Modal from "react-modal";
-import Button from "../Button";
 import Captcha from "../Captcha";
-import CollectionHits from "../CollectionHits";
-import Facet from "../Facet";
 import { FocusError, FormButtons, FormGroup } from "../Form";
-import { CheckBoxInput, YearInput } from "../Inputs";
 import MaterialIcon from "../MaterialIcon";
 import { ModalSavedItemList } from "../SavedItem";
 import "./styles.scss"
 
 
-const MyListModal = (props) => (
+const ModalMyList = (props) => (
   <Modal
     appElement={props.appElement ? props.appElement : Modal.setAppElement("#root")}
     isOpen={props.isOpen}
@@ -37,7 +33,7 @@ const MyListModal = (props) => (
   </Modal>
 )
 
-MyListModal.propTypes = {
+ModalMyList.propTypes = {
   appElement: PropTypes.object,
   handleChange: PropTypes.func,
   isOpen: PropTypes.bool.isRequired,
@@ -47,7 +43,7 @@ MyListModal.propTypes = {
 }
 
 export const EmailModal = props => (
-  <MyListModal
+  <ModalMyList
     appElement={props.appElement}
     title="Email List"
     handleChange={props.handleChange}
@@ -131,7 +127,7 @@ EmailModal.propTypes = {
 }
 
 export const ReadingRoomRequestModal = props => (
-  <MyListModal
+  <ModalMyList
     appElement={props.appElement}
     title="Request in Reading Room"
     handleChange={props.handleChange}
@@ -213,7 +209,7 @@ ReadingRoomRequestModal.propTypes = {
 
 
 export const DuplicationRequestModal = props => (
-  <MyListModal
+  <ModalMyList
     appElement={props.appElement}
     title="Request Copies"
     handleChange={props.handleChange}
@@ -324,128 +320,5 @@ DuplicationRequestModal.propTypes = {
   isOpen: PropTypes.bool.isRequired,
   list: PropTypes.array.isRequired,
   submitList: PropTypes.array.isRequired,
-  toggleModal: PropTypes.func.isRequired,
-}
-
-export const FacetModal = props => {
-  var [startYear, setStartYear] = useState(0);
-  var [endYear, setEndYear] = useState(0);
-
-  useEffect(() => {
-    const startDate = (props.params.start_date__gte ? props.params.start_date__gte : (props.data.min_date && props.data.min_date.value)) || "";
-    const endDate = (props.params.end_date__lte ? props.params.end_date__lte : (props.data.max_date && props.data.max_date.value)) || "";
-    setStartYear(startDate);
-    setEndYear(endDate);
-  }, [props.params.start_date__gte, props.params.end_date__lte, props.data.min_date, props.data.max_date] );
-
-  return (
-    <Modal
-      appElement={props.appElement ? props.appElement : Modal.setAppElement("#root")}
-      isOpen={props.isOpen}
-      onRequestClose={props.toggleModal}
-      className="modal-content--facet"
-      overlayClassName={{
-        base: "modal-overlay--facet slide--right",
-        afterOpen: "slide--right--after-open",
-        beforeClose: "slide--right--before-close"
-      }}
-      closeTimeoutMS={200} >
-      <div className="modal-header--search">
-        <h2 className="modal-header__title">Filter Search Results</h2>
-        <button className="modal-header__button" aria-label="Close" onClick={props.toggleModal}>
-          <MaterialIcon icon="close"/>
-        </button>
-      </div>
-      <div className="modal-body">
-        <Facet>
-          <CheckBoxInput
-            id="online"
-            name="true"
-            className="facet__input"
-            checked={props.params.online === "true"}
-            handleChange={e => props.handleChange(e, "online")}
-            label={`Show me digital materials only (${props.data.online && props.data.online.doc_count})`} />
-        </Facet>
-        <Facet title="Date Range">
-          <YearInput
-            id="startYear"
-            label="Start Year"
-            className="hide-label"
-            handleChange={e => {setStartYear(e.target.value)}}
-            value={startYear} />
-          <YearInput
-            id="endYear"
-            label="End Year"
-            className="hide-label"
-            handleChange={e => {setEndYear(e.target.value)}}
-            value={endYear} />
-          <Button className="btn--sm btn--blue" label="apply" handleClick={() => {props.handleDateChange(startYear, endYear)}}/>
-        </Facet>
-        <Facet
-          handleChange={props.handleChange}
-          items={props.data.format}
-          paramKey="genre"
-          params={props.params.genre}
-          title="Format" />
-        <Facet
-          handleChange={props.handleChange}
-          items={props.data.creator}
-          paramKey="creator"
-          params={props.params.creator}
-          title="Creator" />
-        <Facet
-          handleChange={props.handleChange}
-          items={props.data.subject}
-          paramKey="subject"
-          params={props.params.subject}
-          title="Subject" />
-      </div>
-    </Modal>
-  )
-}
-
-FacetModal.propTypes = {
-  appElement: PropTypes.object,
-  data: PropTypes.object.isRequired,
-  handleChange: PropTypes.func.isRequired,
-  handleDateChange: PropTypes.func.isRequired,
-  isOpen: PropTypes.bool.isRequired,
-  params: PropTypes.object.isRequired,
-  toggleModal: PropTypes.func.isRequired,
-}
-
-export const CollectionHitsModal = props => {
-  return (
-    <Modal
-      appElement={props.appElement ? props.appElement : Modal.setAppElement("#root")}
-      isOpen={props.isOpen}
-      onRequestClose={props.toggleModal}
-      className="modal-content--hits"
-      overlayClassName={{
-        base: "modal-overlay--hits slide--left",
-        afterOpen: "slide--left--after-open",
-        beforeClose: "slide--left--before-close"
-      }}
-      closeTimeoutMS={200} >
-      <div className="modal-header--search">
-        <h2 className="modal-header__title">Inside This Collection</h2>
-        <button className="modal-header__button" aria-label="Close" onClick={props.toggleModal}>
-          <MaterialIcon icon="close"/>
-        </button>
-      </div>
-      <div className="modal-body">
-        <CollectionHits
-          collection={props.data}
-          isLoading={props.isLoading}
-          params={props.params} />
-      </div>
-    </Modal>
-  )
-}
-
-CollectionHitsModal.propTypes = {
-  appElement: PropTypes.object,
-  data: PropTypes.object.isRequired,
-  isOpen: PropTypes.bool.isRequired,
   toggleModal: PropTypes.func.isRequired,
 }
