+++ src/components/ModalSearch/styles.scss
@@ -1,0 +1,2 @@
+@import "../../styles/components/modal-base";
+@import "../../styles/components/modal-search";
