+++ src/styles/components/_modal-base.scss
@@ -53,30 +53,6 @@
   outline: none;
 }
 
-.modal-content--facet {
-  @extend .modal-content;
-  top: 0;
-  left: 0;
-  right: unset;
-  width: 420px;
-  height: 100vh;
-  border-radius: 0px;
-  border: 1px solid $dark-gray;
-  background: $pale-blue;
-}
-
-.modal-content--hits {
-  @extend .modal-content;
-  left: unset;
-  top: 0;
-  right: 0;
-  width: 420px;
-  height: 100vh;
-  border-radius: 0px;
-  border: 1px solid $dark-gray;
-  background: $pure-white;
-}
-
 .modal-header {
   font-family: $sans-serif-stack;
   color: $pure-white;
@@ -120,76 +96,4 @@
   @media screen and (min-width: $break-lg) {
     width: 100%
   }
-}
-
-.modal-list {
-  width: 100%;
-  float: left;
-  padding: 30px 20px;
-  box-sizing: border-box;
-  @media screen and (min-width: $break-lg) {
-    @include grid-column(8);
-  }
-}
-
-.modal-form {
-  float: left;
-  width: 100%;
-  height: 100%;
-  background-color: $off-white;
-  border-top: 1px solid $neutral-sand;
-  padding: 30px 20px;
-  box-sizing: border-box;
-  font-size: 14px;
-  margin-bottom: 18px;
-  @media screen and (min-width: $break-lg) {
-    @include grid-column(4);
-    float: right;
-  }
-}
-
-.modal-form input, textarea, select {
-  width: 100%;
-  border: 1px solid $dark-gray;
-  padding: 10px;
-  box-sizing: border-box;
-  &[type=checkbox] {
-    width: unset;
-    float: left;
-  }
-
-}
-
-.modal-form label {
-  margin-bottom: 5px;
-  margin-top: 10px;
-  float: left;
-  &[for=costs] {
-    float: none;
-    & a {
-      font-family: $sans-serif-stack;
-      font-size: 14px;
-    }
-  }
-}
-
-.modal-form .is-invalid {
-    border: 1px solid $burnt-orange;
-}
-
-.modal-form__buttons {
-  margin: 20px 0;
-}
-
-.modal-form__error {
-  border-radius: 3px;
-  border: 1px solid $burnt-orange;
-  font-family: $sans-serif-stack;
-  font-size: 13px;
-  font-style: italic;
-  display: block;
-  background: $light-orange;
-  color: $dark-gray;
-  margin: 10px 0px;
-  padding: 10px;
 }
