+++ src/styles/components/_modal-search.scss
@@ -1,0 +1,32 @@
+.modal-content--facet {
+  @extend .modal-content;
+  top: 0;
+  left: 0;
+  right: unset;
+  width: 420px;
+  height: 100vh;
+  border-radius: 0px;
+  border: 1px solid $dark-gray;
+  background: $pale-blue;
+}
+
+.modal-content--hits {
+  @extend .modal-content;
+  left: unset;
+  top: 0;
+  right: 0;
+  width: 420px;
+  height: 100vh;
+  border-radius: 0px;
+  border: 1px solid $dark-gray;
+  background: $pure-white;
+}
+
+.modal-header--search {
+  color: $deep-navy;
+  background-color: none;
+  padding: 30px 30px 40px 30px;
+  float: left;
+  width: 100%;
+  box-sizing: border-box;
+}
